requirements.txt
fastapi
uvicorn
pyyaml
openai
gitpython